refactor : 텍스트 퀴즈 리팩토링 완료
